Move ConferenceTimeFrame/Test into "models" package.
