+++ app/src/main/java/nerd/tuxmobil/fahrplan/congress/MyApp.kt
@@ -6,8 +6,8 @@ import androidx.annotation.VisibleForTesting
 import info.metadude.android.eventfahrplan.commons.logging.Logging
 import info.metadude.android.eventfahrplan.commons.temporal.Moment
 import info.metadude.android.eventfahrplan.commons.temporal.Moment.Companion.toMoment
+import nerd.tuxmobil.fahrplan.congress.models.ConferenceTimeFrame
 import nerd.tuxmobil.fahrplan.congress.repositories.AppRepository
-import nerd.tuxmobil.fahrplan.congress.utils.ConferenceTimeFrame
 import org.ligi.tracedroid.TraceDroid
 import org.threeten.bp.LocalDate
 import org.threeten.bp.LocalTime
